+++ R/data.R
@@ -21,15 +21,12 @@
 ##' We should have a brief description here
 ##'
 ##' @format ## `aart`
-##' A data frame with 47,315 rows and 3 columns:
+##' A data frame with 1,553 rows and 2 columns:
 ##' \describe{
-##'
-##' \item{\code{risk_cat}}{Credit score category}
-##'   \item{\code{Zi}}{Time-of-event}
+##'   \item{\code{Xi}}{Time to event}
 ##'   \item{\code{Yi}}{Truncation time}
-##'   \item{\code{Di}}{Default indicator (1 = default, 0 = not default)}
 ##' }
-##' @source DATASETS SOURCE (I'll get from the paper)
+##' @source TBD 
 "aart"
 
 ##' NAME OF THE DATASET?
@@ -39,10 +36,9 @@
 ##' @format ## `mbalt`
 ##' A data frame with 1,553 rows and 2 columns:
 ##' \describe{
-##'
-##' \item{\code{risk_cat}}{Credit score category}
-##'   \item{\code{Xi}}{Time-of-event}
+##'   \item{\code{Zi}}{Time to event}
 ##'   \item{\code{Yi}}{Truncation time}
+##'   \item{\code{Di}}{Default indicator}
 ##' }
 ##' @source DATASETS SOURCE (I'll get from the paper)
 "mbalt"